added firmware update function
