+++ src/cloud/IoTalkie.MessageSenderCmd/AzureSettings.cs
@@ -5,6 +5,9 @@
         public string IoTHubConnectionString { get; set; }
 
         public string BlobStorageConnectionString { get; set; }
+        
         public string AudioMessagesContainerName { get; set; }
+        
+        public string FirmwareUpdatesContainerName { get; set; }
     }
 }
