+++ src/cloud/IoTalkie.MessageSenderCmd/Program.cs
@@ -1,15 +1,9 @@
-﻿using Azure.Storage.Blobs;
-using Azure.Storage.Sas;
-using Microsoft.Azure.Devices;
-using Microsoft.Extensions.Configuration;
+﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using System;
 using System.CommandLine;
 using System.CommandLine.Invocation;
-using System.IO;
-using System.Text;
-using System.Text.Json;
 using System.Threading.Tasks;
 
 namespace IoTalkie.MessageSenderCmd
@@ -28,6 +22,8 @@ namespace IoTalkie.MessageSenderCmd
 
             services.Configure<AzureSettings>(configuration.GetSection(typeof(AzureSettings).Name));
             services.AddSingleton<SendMessageFunctionality>();
+            services.AddSingleton<UpdateFirmwareFunctionality>();
+
             var provider = services.BuildServiceProvider();
             return provider;
         }
@@ -50,7 +46,16 @@ namespace IoTalkie.MessageSenderCmd
             };
 
             sendMessage.Handler = CommandHandler.Create<string, string, string, IConsole>(SendMessageAsync);
-            var cmd = new RootCommand { sendMessage };
+
+            Command updateFirmware = new Command("updateFirmware", "Upload firmware and advice a single device download to update.")
+            {
+                new Argument<string>("file", "The file to be sent."),
+                new Argument<string>("deviceId", "Target device the message should be sent to"),
+            };
+
+            updateFirmware.Handler = CommandHandler.Create<string, string, IConsole>(UpdateFirmwareAsync);
+
+            var cmd = new RootCommand { sendMessage, updateFirmware };
 
             return cmd.Invoke(args);
         }
@@ -58,6 +63,10 @@ namespace IoTalkie.MessageSenderCmd
         private static async Task SendMessageAsync(string file, string deviceId, string senderId, IConsole arg4)
         {
             await services.GetService<SendMessageFunctionality>().Run(file, deviceId, senderId);
+        }
+        private static async Task UpdateFirmwareAsync(string file, string deviceId, IConsole arg4)
+        {
+            await services.GetService<UpdateFirmwareFunctionality>().Run(file, deviceId);
         }
     }
 
@@ -76,71 +85,18 @@ namespace IoTalkie.MessageSenderCmd
         public string RemoteUrl { get; set; }
     }
 
-    class SendMessageFunctionality
+    public class UpdateFirmwareCmd
     {
-        private readonly AzureSettings settings;
+        public string Cmd { get; set; } = "updateFirmware";
 
-        public SendMessageFunctionality(IOptions<AzureSettings> provider)
-        {
-            this.settings = provider.Value;
-        }
+        public string MessageId { get; set; } = Guid.NewGuid().ToString();
 
-        internal async Task Run(string localFilePath, string deviceId, string senderId)
-        {
-            if (!File.Exists(localFilePath))
-            {
-                Console.WriteLine("File not found");
-                return;
-            }
-
-            var fileName = Path.GetFileName(localFilePath);
-
-            // Create a BlobServiceClient object which will be used to create a container client
-            BlobServiceClient blobServiceClient = new BlobServiceClient(settings.BlobStorageConnectionString);
-
-            // Create the container and return a container client object
-            BlobContainerClient containerClient = blobServiceClient.GetBlobContainerClient(settings.AudioMessagesContainerName);
-
-            // Get a reference to a blob
-            BlobClient blobClient = containerClient.GetBlobClient(fileName);
-
-            Console.WriteLine("Uploading to Blob storage as blob:\n\t {0}\n", blobClient.Uri);
-
-            // Open the file and upload its data
-            using FileStream uploadFileStream = File.OpenRead(localFilePath);
-            await blobClient.UploadAsync(uploadFileStream, true);
-            uploadFileStream.Close();
-
-            var sasBuilder = new BlobSasBuilder(BlobSasPermissions.Read, expiresOn: DateTimeOffset.UtcNow.AddDays(60))
-            {
-                StartsOn = DateTimeOffset.UtcNow.AddMinutes(-5),
-                BlobContainerName = settings.AudioMessagesContainerName,
-                BlobName = fileName
-            };
-
-            var downloadUrl = blobClient.GenerateSasUri(sasBuilder);
-
-            Console.WriteLine($"Download Url (with sas token): {downloadUrl}");
-
-            ServiceClient service = ServiceClient.CreateFromConnectionString(settings.IoTHubConnectionString);
-            
-            var ms = new MemoryStream();
-            var jsonMessage = new NewMessageCmd { SenderId = senderId, RemoteUrl = downloadUrl.ToString(), Size = new FileInfo(localFilePath).Length };
-            var options = new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.CamelCase, Encoder = System.Text.Encodings.Web.JavaScriptEncoder.UnsafeRelaxedJsonEscaping };
-            var jsonString = JsonSerializer.Serialize(jsonMessage, options);
-
-            Console.WriteLine("Sending message: " + jsonString);
-
-            // new Message(Encoding.ASCII.GetBytes(ms.ToArray());
+        public long Timestamp { get; set; } = DateTimeOffset.Now.ToUnixTimeSeconds();
 
 
-            Message message = new Message(Encoding.ASCII.GetBytes(jsonString));
-            message.ExpiryTimeUtc = DateTime.UtcNow.AddHours(8);
+        public long Size { get; set; }
 
-            await service.SendAsync(deviceId, message);
-            
-            Console.WriteLine("Message sent");
-
-        }
+        public string RemoteUrl { get; set; }
     }
+
 }
